ES-1 - rename files and refactor methods
